Changes to survey index page while fixing rspec
